requirements.txt
streamlit>=1.36.0
openai>=1.60.0
supabase>=2.10.0
python-dotenv>=1.0.1
pydantic>=2.8.0
pydantic-settings>=2.4.0
tenacity>=8.5.0
python-dateutil>=2.9.0
pytest>=8.0.0
